PythonMySQL-main/requirements.txt

aws-cfn-bootstrap==2.0
click==8.0.1
docutils==0.14
Flask==2.0.1
Flask-Cors==3.0.10
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
lockfile==0.11.0
MarkupSafe==2.0.1
mysql-connector==2.2.9
pystache==0.5.4
python-daemon==2.2.3
simplejson==3.2.0
six==1.16.0
typing-extensions==3.10.0.2
Werkzeug==2.0.1
zipp==3.5.0
